feat: create shiftable module and move shift related methods to it
