+++ lib/key_generator.rb
@@ -1,5 +1,3 @@
-require './lib/enigma'
-
 module KeyGenerator
   def generate_key
     number_array = ["0","1","2","3","4","5","6","7","8","9"]
